[Refactor] AiProcessor 클래스 네임 변경 및 API 호출 로직 AiClient 로 이전.
